pyproject.toml
[project]
name = "lapleaderboard"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "firebase-admin>=6.9.0",
    "flask>=3.0.3",
    "google-auth>=2.40.3",
    "google-cloud-firestore>=2.21.0",
    "google-cloud-logging>=3.12.1",
    "gunicorn>=23.0.0",
    "requests>=2.32.4",
]
